fix(renderer): widen batch indices u16 -> u32 to stop silent overflow (RC5)
All batched 2D content (sprites, UI, text, spine, tilemap, particles) streams
into a single shared Batch vertex/index buffer. Indices were 16-bit, so once a
frame accumulated more than 65,535 batched vertices every `baseVertex` cast
truncated and the GPU read garbage geometry — a silent corruption that triggers
on any moderately busy scene (the audit's #1 latent renderer bug).

- TransientBufferPool: index_staging std::vector<u16> -> u32; writeIndices /
  appendIndices take const u32*; EBO sizing uses sizeof(u32). The const u32*
  signature makes every index-producing call site a compile error until widened.
- Draw calls that read the pool EBO now use GfxDataType::UnsignedInt with a
  sizeof(u32) element stride: DrawList (the merged batch draw) and
  ShapePlugin::customDraw.
- Widen every index array / base-vertex cast / index vector / quad-index
  constant to u32 across all batched plugins and RenderFrameSubmit. Non-index
  u16 (tile ids, text lengths, UTF-8 decode) left unchanged.

Verified: emcc 5.0.0 web build compiles clean and links esengine-core.wasm. The
u32 signature surfaced the one site outside plugins/ (RenderFrameSubmit) that a
manual sweep missed. SpinePlugin is excluded from the spine-off build and was
inspection-verified (identical mechanical change).

+++ src/esengine/renderer/plugins/ShapePlugin.hpp
@@ -28,7 +28,7 @@ private:
         f32 shapeType, halfW, halfH, cornerRadius;
     };
 
-    static constexpr u16 QUAD_INDICES[6] = { 0, 1, 2, 2, 3, 0 };
+    static constexpr u32 QUAD_INDICES[6] = { 0, 1, 2, 2, 3, 0 };
 
     resource::ShaderHandle shape_shader_handle_;
     u32 shape_shader_id_ = 0;
